Remove file from MRU list if open fails
If the user selects a file from the MRU list, and the file fails to
open, then it will now be removed from the list.

Removing the code that reads/writes persistentStore out of
fileReducer.js and into actions/files.js instead. Reducers should be
pure functions without side effects.