Convert to new architecture
Adding `useRegisterDragEvents` and `useLoadSettingsInitially` to
ControlPanel.jsx are a bit of a hack.

`useRegisterDragEvents` should probably be rather in index.jsx but there
currently is no clean way to achieve that.

`useLoadSettingsInitially` should probably be rewritten so that the
initial settings are loaded by the reducers from the store.

But for now, this should work.