use comName to located the serialport
On Windows the .path attribute is mapped to a com port using a string,
to make it more readible it's preferable to use the .comName attribute.

On Linux the .path attribute is equal to the .comName attribute. Which
means it is favorable to use .comName to make it consistent with Windows.