Inline obsolete portPath
The old portPath function was (as described in the comment) just needed
for the transition from serialport v7 to v8.

With the switch from pc-nrfjprog-js to nrf-device-lib we mostly do not
use the serialport anymore. Here we also do not use it and so we do not
need to care about the changed properties. So it is easier to just
inline portPath.

And along the way we can also change the previous
    port ? port.comName : undefined
to
    port?.comName ?? undefined