Fix calls to watch devices
Calling `startWatchingDevices` was wrong in two ways:
- It is a thunk, so the result needs to be dispatched to do anything.
  In the current for it was just a no-op.
- It needs two parameters: the device listing configuration and a
  callback to deselect the current device. But at least the latter is
  not trivial to provide by an app.

So, since the call effectively did nothing, we seemingly can also remove
it. But it is strange to to then leave the call to `stopWatchingDevices`
in there. I also removed it and could not detect any problems that
come out of this change.