Refactor colour algorithm
The old usage of an enum was problematic. And the new is shorter anyhow.

It is intentionally that in the old enum and the new record the values
for MBR_OR_APP are different: In the function `getMBRRegion()` the
colour MBR was used for MBR_OR_APP, so the new algorithm keeps the same
colours for that case.